ap-madang/requirements.txt
asgiref==3.4.1
attrs==21.2.0
awsebcli==3.20.2
bcrypt==3.2.0
blessed==1.19.0
boto3==1.18.64
botocore==1.21.64
cached-property==1.5.2
cement==2.8.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
Django==3.2.7
django-admin-json-editor==0.2.3
django-cors-headers==3.10.0
django-crontab==0.7.1
django-dotenv==1.4.2
django-rest-swagger==2.2.0
django-storages==1.12.2
djangorestframework==3.12.4
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.20.0
future==0.16.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
jmespath==0.10.0
jsonschema==3.2.0
MarkupSafe==2.0.1
mysqlclient==2.0.3
openapi-codec==1.3.2
packaging==21.0
paramiko==2.7.2
pathspec==0.5.9
Pillow==8.4.0
protobuf==3.17.3
pycparser==2.20
PyJWT==2.2.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
semantic-version==2.8.5
sentry-sdk==1.4.3
setuptools-rust==0.12.1
simplejson==3.17.5
six==1.14.0
sqlparse==0.4.2
termcolor==1.1.0
texttable==1.6.4
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.7
wcwidth==0.1.9
websocket-client==0.59.0
